Refactor ManageTurf component: streamline state management, enhance search and filter functionality, and improve UI/UX with new styling and modal for turf updates.
