+++ .firebase/hosting.YnVpbGQ.cache
@@ -1,28 +1,28 @@
 robots.txt,1763358953829,b2090cf9761ef60aa06e4fab97679bd43dfa5e5df073701ead5879d7c68f1ec5
 manifest.json,1769491707545,591318a2dbe25b9e8c0526dabadf31f0d92ce3654dc2341d27db08c96514b851
-index.html,1769803180412,437fbc8918ca42d8db67dbf7d6b44587bde613afb6469f3e39d8c0a4c5a61c7e
-asset-manifest.json,1769803180418,09313fb47fd3c70a3005bda2ba23ca789685ec5d766b283345ce02652bf1de1b
-static/js/main.ee9683a0.js.LICENSE.txt,1769803180445,49f15e87690e86c62d2129a3f11da19e11881a2f4ab254b6a1665910abe3f800
-static/media/Google_Pay_Logo.svg.fdf6d9f103d8af215c7e.png,1769803180408,ff35fdec20eb75a77a79d1972a9705de9ae4f25adff24534a5020b904ca63e57
-static/media/balloon.ca92990e536160406185.TTF,1769803180402,34429138c4e0bd041b952a7035ba1d555aef926bc7288bb99c95233d55a90f57
-static/media/BgImg4.83582a168af655004c08.jpg,1769803180408,e5d7869a406436086f3f026d7fb1863bb4639ae9baee29638a150588b34dc932
-static/media/BgImg3.e088f77d2ccab018a611.jpg,1769803180445,1421269dfdb9847dbaf01336779be32b64a8359dddec0c2a79b29346cdd446c5
-static/media/football.963146bc9639876dc8d3.png,1769803180408,882de7550168110b8d2aa037214971c805a7e2c0d1cfcf87e923c56a9a8b8cb4
-static/media/login_background.0fed8c7bb9de0859aceb.jpeg,1769803180408,134fb271a6453aaeb3526707099aaceab8733ea0907e7b1151cdec46de01416f
-static/media/PickleImg.9a356ccfdcb0e1d2ccc2.png,1769803180408,a1c0bf927238e6eac677c775551fcceb9372a66a8c024bf0be365efa47ae79ac
-static/media/badminton.d10a47906e83e7a3c8c6.png,1769803180408,fa11ecb65bbce15815593c87f1c834b15601459590e9a8f0627b28f0354a5e85
 logo-512.png,1769491707545,0c5b543913c2c5d20411c907c368ee0cc1514106465f94e0ca178337ddb41d7a
-static/media/AboutImg.2c1d28afbf1625a7b764.jpg,1769803180408,ecec563996a6df901098cdd178b8cd4dbc08525df336cf9d6fc731c2111eb530
 logo-192.png,1769491707543,0c5b543913c2c5d20411c907c368ee0cc1514106465f94e0ca178337ddb41d7a
 favicon.png,1769491707539,0c5b543913c2c5d20411c907c368ee0cc1514106465f94e0ca178337ddb41d7a
-static/media/boxcricket_football.e4df7d8149dd67016060.png,1769803180408,f230fc86733339d0bc642cdf3c9eb8286a47f8080a1f48f07845f5d78cbdcfaa
-static/media/bootstrap-icons.92ea18a81d737146ff04.woff2,1769803180408,9bbd1a9a4f2c1815120516b2f5df05a97e5078c62268e2f6a2a36d8d54b6b725
-static/css/main.bb592534.css,1769803180445,c3bc5b7a71f49bfd37674e69ea2279d0386b14b6e27642aacee5259b57e5cd71
-static/media/bootstrap-icons.1295669cd4e305c97f2c.woff,1769803180402,4a0b4e72f5538b3f1a654d4a54031127a6337e84e8664502dbdeb3abcdc326d6
-static/media/BgImg1.79344000d28630410e3f.jpg,1769803180408,3437fe1d4e01fdf380fb40911c3d2dbe75468dbecb909b1c6a07da211ae04299
-static/media/LogoImg.f6ee4f5735264ef97a84.png,1769803180408,b4a0f00b6ba198d596056b37f8a410d396e4796c2eda607f2eb630a065e72895
-static/media/BgImg2.0526b66777362801a7d0.jpg,1769803180408,5e2c735ab32f71b7a82a6dde32010f87a7725db314f4f44dd4f79827f2d9c85b
-static/media/BgImg.22f26f893e6f4c7fc055.jpeg,1769803180408,eafe0675afdd4fc5b3f028b37852fc2147fc173b6073fea7428a2a5e04d5d62a
-static/css/main.bb592534.css.map,1769803180445,7df39c6e7fa5ca2a8b774a52f68a9b4d4c79df26c971839a7b53de74107adaa9
-static/js/main.ee9683a0.js,1769803180445,43cca5503e5544173b77d778cd588577ec161c2a4c5a5a8386a139f930519ec8
-static/js/main.ee9683a0.js.map,1769803180452,b6479e8510b63e949c02ed18b45386c9da8627a2eca5b341ee5da69b25db3aae
+index.html,1770036988958,16e51360b86c6a108ab55f510cb1972728aec52bf56198fbee2c1e70ab112e47
+asset-manifest.json,1770036988980,69f62caa137b2f4d8bfa9591da2c34a3541dc9e87c4ede0f708e3e2c95bfae46
+static/js/main.cef8b04d.js.LICENSE.txt,1770036988980,4ccb18306f5be81bacd9e6ffd71d3d986690abbb60e229919f369cb50b8b81bb
+static/media/Google_Pay_Logo.svg.fdf6d9f103d8af215c7e.png,1770036988948,ff35fdec20eb75a77a79d1972a9705de9ae4f25adff24534a5020b904ca63e57
+static/media/balloon.ca92990e536160406185.TTF,1770036988946,34429138c4e0bd041b952a7035ba1d555aef926bc7288bb99c95233d55a90f57
+static/media/football.963146bc9639876dc8d3.png,1770036988948,882de7550168110b8d2aa037214971c805a7e2c0d1cfcf87e923c56a9a8b8cb4
+static/media/BgImg4.83582a168af655004c08.jpg,1770036988948,e5d7869a406436086f3f026d7fb1863bb4639ae9baee29638a150588b34dc932
+static/media/BgImg3.e088f77d2ccab018a611.jpg,1770036988980,1421269dfdb9847dbaf01336779be32b64a8359dddec0c2a79b29346cdd446c5
+static/media/login_background.0fed8c7bb9de0859aceb.jpeg,1770036988950,134fb271a6453aaeb3526707099aaceab8733ea0907e7b1151cdec46de01416f
+static/media/PickleImg.9a356ccfdcb0e1d2ccc2.png,1770036988946,a1c0bf927238e6eac677c775551fcceb9372a66a8c024bf0be365efa47ae79ac
+static/media/badminton.d10a47906e83e7a3c8c6.png,1770036988950,fa11ecb65bbce15815593c87f1c834b15601459590e9a8f0627b28f0354a5e85
+static/media/AboutImg.2c1d28afbf1625a7b764.jpg,1770036988946,ecec563996a6df901098cdd178b8cd4dbc08525df336cf9d6fc731c2111eb530
+static/media/boxcricket_football.e4df7d8149dd67016060.png,1770036988948,f230fc86733339d0bc642cdf3c9eb8286a47f8080a1f48f07845f5d78cbdcfaa
+static/media/bootstrap-icons.92ea18a81d737146ff04.woff2,1770036988946,9bbd1a9a4f2c1815120516b2f5df05a97e5078c62268e2f6a2a36d8d54b6b725
+static/media/bootstrap-icons.1295669cd4e305c97f2c.woff,1770036988946,4a0b4e72f5538b3f1a654d4a54031127a6337e84e8664502dbdeb3abcdc326d6
+static/media/BgImg1.79344000d28630410e3f.jpg,1770036988950,3437fe1d4e01fdf380fb40911c3d2dbe75468dbecb909b1c6a07da211ae04299
+static/css/main.d76fb6cf.css,1770036988980,30882fc13e6862a262743c7fda47f665a024e4663412fe3360fd47cbc79ad208
+static/media/LogoImg.f6ee4f5735264ef97a84.png,1770036988950,b4a0f00b6ba198d596056b37f8a410d396e4796c2eda607f2eb630a065e72895
+static/media/BgImg2.0526b66777362801a7d0.jpg,1770036988950,5e2c735ab32f71b7a82a6dde32010f87a7725db314f4f44dd4f79827f2d9c85b
+static/media/BgImg.22f26f893e6f4c7fc055.jpeg,1770036988950,eafe0675afdd4fc5b3f028b37852fc2147fc173b6073fea7428a2a5e04d5d62a
+static/css/main.d76fb6cf.css.map,1770036988980,f18a030ecc3ca072566981881f12ea70a7f5dd59ab53f454e60e3c940858ee38
+static/js/main.cef8b04d.js,1770036988980,3b3de6db2a1fbf347e35164f5da98ff3472f0dc1c045bb4ea23c10307bab0e9c
+static/js/main.cef8b04d.js.map,1770036988985,9cf0c01138d9f4fc6ec205aea635dde5fa48d44b34f7b7a94be5e3bd1a1260d8
